Rename api module as routes
