+++ api/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod api;
 pub mod conf;
 pub mod db;
 pub mod models;
+pub mod routes;
