+++ api/src/routes/catalogs.rs
