+++ api/src/routes/query.rs
